feat: Refactor folder-filters-menu into modular architecture
🏗️ MAJOR REFACTORING: Split 779-line monolithic component into maintainable modules

## 📊 Before vs After
- **Before**: 1 file (779 lines) - difficult to maintain and navigate
- **After**: 15 files (30-200 lines each) - modular, focused, maintainable

## 🔧 New Structure

## ✅ Benefits
- **Maintainability**: Single responsibility per module
- **Readability**: Focused, easy-to-understand files
- **Reusability**: Sections can be reused in other contexts
- **Testability**: Individual modules easier to test
- **Collaboration**: Teams can work on different sections
- **Performance**: Potential for lazy loading

## 🔄 Backward Compatibility
- All existing imports continue to work
- No breaking changes to external API
- Compatibility layer maintains existing functionality

## 🧪 Validation
- TypeScript compilation successful
- No circular dependencies
- All filter functionality preserved
- Modular architecture fully implemented
